awslambda/requirements.txt
google-auth==1.2.1
python_dateutil==2.6.1
pytz==2017.3
requests==2.18.4
